Added: finding first position float maximum number

+++ src/main/java/com/bridgelabz/generic/MaxNumber.java
@@ -1,9 +1,8 @@
 package com.bridgelabz.generic;
 
 public class MaxNumber {
-    public int getMaxNum(Integer num1, Integer num2, Integer num3) {
-        Integer max = num1;
-      //  System.out.println("max number is out of if>>"+max);
+    public float getMaxNum(Float num1, Float num2, Float num3) {
+        float max = num1;
         if (num2.compareTo(max)>0)
             max = num2;
         if (num3.compareTo(max)>0)
